perf(boot): split init into phased pipeline — Ping-ready in <1ms, defer network/zram to background
Restructure VM boot across tiny-init, guest-agent, and host to minimize
time-to-first-Ping.

tiny-init:
- Defer sysctl hardening, /dev/shm, /dev/fd symlinks, and zram device
  setup to guest-agent — only keep modules_disabled=1 and module loading
- Parallelize /dev/vda + virtio-ports wait (B2)
- Tighter device polling backoff: 31.7ms max vs 63ms (B3)
- Add boot timing instrumentation (monotonic_us)

guest-agent phased init:
- Phase 1: env vars only (<1ms) → start listening immediately
- Phase 2 core: chmod (libc, no fork), mounts, sysctl, dev symlinks,
  /dev/shm, pre-write REPL wrappers (~10ms)
- Background: network setup + zram swap (non-blocking)
- Gate ExecuteCode/InstallPackages on NETWORK_READY; Ping/file I/O
  respond instantly without waiting
- Quiet mode (init.quiet=1): suppress eprintln! to avoid MMIO/UART
  trap overhead on critical path (E4)

host (qemu_cmd.py, vm_manager.py):
- Kernel cmdline: quiet, loglevel=0, init_on_alloc=0, no_timer_check,
  scsi_mod.scan=none, audit=0, edd=off, io_delay=none
- virtio-blk event_idx=off, balloon free-page-reporting=off during boot
- Tighter guest detection backoff (max 50ms intervals vs 1s)
- timeout_seconds: int → float for sub-second granularity

+++ src/exec_sandbox/guest_channel.py
@@ -123,7 +123,7 @@ class GuestChannel(Protocol):
     Uses structural typing (Protocol) instead of inheritance.
     """
 
-    async def connect(self, timeout_seconds: int) -> None:
+    async def connect(self, timeout_seconds: float) -> None:
         """Establish connection to guest agent."""
         ...
 
@@ -253,7 +253,7 @@ class UnixSocketChannel:
         self._write_task: asyncio.Task[None] | None = None
         self._shutdown_event: asyncio.Event = asyncio.Event()
 
-    async def connect(self, timeout_seconds: int) -> None:
+    async def connect(self, timeout_seconds: float) -> None:
         """Connect to guest via Unix socket with mandatory peer verification.
 
         Single connection attempt with timeout (no retry).
@@ -578,7 +578,7 @@ class DualPortChannel:
         self._event_channel: UnixSocketChannel = UnixSocketChannel(event_socket, expected_uid)
         self._dispatcher: FileOpDispatcher | None = None
 
-    async def connect(self, timeout_seconds: int) -> None:
+    async def connect(self, timeout_seconds: float) -> None:
         """Connect both command and event ports.
 
         Connects in parallel for speed. Single connection attempt with timeout (no retry).
